Update bpf.h
Update as of commit fcf85729d8ef ("Merge branch 'fix-bpf-helpers-doc'").
This brings fixes for existing documentation (where kernel structs were
used instead of user structs), and documentation for a new helper.

The documentation for the new helper is patched, because it is not
correctly formatted in kernel version of bpf.h. Another patch to be
sent.

Also add a `all` default target to Makefile because both man and HTML
pages are important in this repo.